Desarrollo/APP_INVENTARIO/inventario/wils0n-eshop-3f08e2e976f3/requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

# Python-PostgreSQL Database Adapter
# If using Win for dev, this assumes Unix in prod
# ------------------------------------------------
psycopg2==2.6.1

# WSGI Handler
# ------------------------------------------------
gevent==1.1.1
gunicorn==19.5.0

# Static and Media Storage
# ------------------------------------------------
boto==2.40.0
django-storages-redux==1.3.2


# Mailgun Support
# ---------------
django-mailgun==0.9.1






